apps/identity-service/requirements.txt
# Django 5.2 LTS - Current stable (NOT 6.0)
Django>=5.2,<5.3
graphene-django>=3.2.0
graphene-federation>=3.3.0

# Database
psycopg[binary]>=3.2.0

# Kafka
confluent-kafka>=2.6.0

# Server
gunicorn>=23.0.0

# Utilities
python-dotenv>=1.0.0

# Authentication
PyJWT>=2.9.0

# Testing
pytest>=8.0.0
pytest-django>=4.8.0
pytest-mock>=3.14.0
djangorestframework>=3.15.0
django-cors-headers>=4.6.0
